% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralapp-package.R
\docType{package}
\name{admiralapp-package}
\alias{admiralapp}
\alias{admiralapp-package}
\title{admiralapp: An App to Explore and Discover Documentation}
\description{
An app to explore and discover documentation of the toy admiralverse.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
