% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\docType{data}
\name{public}
\alias{public}
\title{A dataframe with the documentation of the toy 'admiralverse'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 399 rows and 4 columns.
}
\usage{
public
}
\description{
See also \url{https://maurolepore.github.io/admiralverse/}.
}
\keyword{datasets}
